feat: change attributes of entities that represents money values from number to Money data type. And refact tests about it.

+++ src/infra/database/knex/migrations/20220304193739_create_voucher.ts
@@ -8,7 +8,7 @@ export async function up(knex: Knex): Promise<void> {
         table.string('code').notNullable().unique();
         table.string('type').notNullable();
         table.integer('amount').notNullable();
-        table.float('minValue');
+        table.integer('minValue');
         table.timestamp('created_at').defaultTo(knex.fn.now())
         table.timestamp('updated_at').defaultTo(knex.fn.now())
     })
